requirements.txt
dnspython
python-docx